Rename ConfirmModel.tsx to fix typo
